+++ src/components/report/Tap.js
@@ -19,30 +19,30 @@ function Tap({ leftTap, setLeftState, setRightState }) {
 
   return (
     <TapWrapper>
-          {/* 탭버튼 */}
-            {leftTap ? (
-              // 내탭 클릭(초기상태)
-            <TapBtnWrapper>
-              <TapBtn1>나의 증상분석</TapBtn1>
-              <TapBtn2 onClick={toggleTap}>짝꿍의 증상분석</TapBtn2>
-            </TapBtnWrapper>
-            )
-            :(
-              // 짝꿍탭 클릭
-            <TapBtnWrapper>
-              <TapBtn2 onClick={toggleTap}>나의 증상분석</TapBtn2>
-              <TapBtn1>짝꿍의 증상분석</TapBtn1>
-            </TapBtnWrapper>
-            )}
+      {/* 탭버튼 */}
+        {leftTap ? (
+          // 내탭 클릭(초기상태)
+        <TapBtnWrapper>
+          <TapBtn1>나의 증상분석</TapBtn1>
+          <TapBtn2 onClick={toggleTap}>짝꿍의 증상분석</TapBtn2>
+        </TapBtnWrapper>
+        )
+        :(
+          // 짝꿍탭 클릭
+        <TapBtnWrapper>
+          <TapBtn2 onClick={toggleTap}>나의 증상분석</TapBtn2>
+          <TapBtn1>짝꿍의 증상분석</TapBtn1>
+        </TapBtnWrapper>
+        )}
 
-          {/* 그래프*/}
-          <GraphWrapper>
-            <LineChart />
-          </GraphWrapper>
+      {/* 그래프*/}
+      <GraphWrapper>
+        <LineChart />
+      </GraphWrapper>
 
-          {/* 슬라이드*/}
-          <SlideAnimation />
-        </TapWrapper>
+      {/* 슬라이드*/}
+      <SlideAnimation />
+    </TapWrapper>
   )
 }
